requirements.txt
Flask==2.1.2
flask_bcrypt==1.0.1
flask_login==0.6.1
flask_mail==0.9.1
flask_sqlalchemy==2.5.1
flask_wtf==1.0.1
itsdangerous==2.1.2
Pillow==9.1.1
WTForms==3.0.1
pytest==7.1.2
